refactor(team): split spawn orchestration off pane_lifecycle.go (C24)
pane_lifecycle.go grew to 893 lines across the C5b-C5f series. It
bundled three distinct concerns: tmux primitive wrappers (single-call
methods like SplitFirstAgent, NewOverflowWindow, ApplyMainVerticalLayout),
high-level spawn orchestration (multi-step flows like SpawnVisibleAgents
that compose those primitives), and pure helpers (channelPaneNeedsRespawn,
isMissingTmuxSession, etc.).

Splitting on the orchestration boundary gives each file a sharp single
responsibility and matches the file-naming pattern used elsewhere in
the package (headless_codex.go vs headless_codex_runner.go,
launcher_lifecycle.go before C20's three-way split).

* pane_lifecycle.go (596 lines, was 893) — paneLifecycleDeps struct,
  paneLifecycle type + constructors + withClock test seam, the 23
  tmux primitive methods (HasLiveSession, ListTeamPanes, SplitFirstAgent,
  ApplyMainVerticalLayout, IsPaneDead, ...), and the pure helpers
  (parseAgentPaneIndices, paneFallbackMessages, etc.).

* pane_lifecycle_spawn.go (313 lines, new) — the seven high-level
  orchestration methods that compose tmux primitives into roster-aware
  flows: SpawnVisibleAgents, SpawnOverflowAgents,
  DetectDeadPanesAfterSpawn, TrySpawnWebAgentPanes, PrimeVisibleAgents,
  ReportPaneFallback, CaptureDeadChannelPane.

The split follows the same logic as launcher.go's "tmux primitives stay,
high-level orchestration moves" — the primitives are stable building
blocks; orchestrations evolve as the spawn flow does.

Stacked on refactor/launcher-c23-headless-helpers (C23).
